requirements.txt
# Core dependencies
sentence-transformers==2.7.0
faiss-cpu==1.8.0
rank-bm25==0.2.2
python-dotenv==1.0.1
groq==0.11.0

# Data collection
beautifulsoup4==4.12.3
requests==2.31.0
lxml==5.2.1

# Text processing
langchain-text-splitters==0.2.0

# UI
streamlit==1.33.0

# Utilities
numpy==1.26.4
pandas==2.2.1
tqdm==4.66.2
